update Map
added gps loctaion to map view controller and google map view

Now the user can see there location on a map where they go
