+++ Dungeons and Dieting/Dungeons and Dieting/GameViewController.swift
@@ -9,18 +9,12 @@
 import UIKit
 import SpriteKit
 import GameplayKit
-import CoreLocation
 
-class GameViewController: UIViewController, CLLocationManagerDelegate {
+class GameViewController: UIViewController{
 
     // Changed Nothing
     
     var playButton: UIButton
-    let LocationM = CLLocationManager()
-    let Clientid = "WP0KU1PZ5EZELDOTL2MVVWEK0Y4PSU45TMHLHUCV4QFBK4OI"
-    let CLIENTSECRET = "FJJPDA5ANL3JCXAUALOMXKLWJYKSUY1PTZXESPGYO04RA1RK"
-
-    
         
     init() {
         playButton = UIButton(frame: CGRect (x: 0, y: 0, width: 100, height: 100))
@@ -43,7 +37,6 @@ class GameViewController: UIViewController, CLLocationManagerDelegate {
         playButton.centerYAnchor.constraint(equalTo: view.centerYAnchor).isActive = true
         playButton.centerXAnchor.constraint(equalTo: view.centerXAnchor).isActive = true
         
-          setupLoation()
     }
     
     @IBAction func toMap() {
@@ -61,56 +54,6 @@ class GameViewController: UIViewController, CLLocationManagerDelegate {
     }
 
   
-    
-    func setupLoation()
-       {
-           LocationM.delegate = self
-           LocationM.desiredAccuracy = kCLLocationAccuracyHundredMeters
-           LocationM.requestWhenInUseAuthorization()
-           LocationM.requestLocation()
-           LocationM.startUpdatingLocation()
-       }
-    
-    
-    func locationManager(_ manager: CLLocationManager, didUpdateLocations locations: [CLLocation])
-    {
-        if let location =  locations.last
-        {
-            if location.horizontalAccuracy > 0 {
-                LocationM.stopUpdatingLocation()
-                
-                let longitude =  location.coordinate.longitude
-                let latitude  = location.coordinate.latitude
-                print("long =\(longitude),latiude= \(latitude )" )
-            }
-        }
-        
 
-        let session = URLSession.shared
-        guard let url = URL(string: "https://api.foursquare.com/v2/venues/search?ll=\(locations[0].coordinate.latitude),\(locations[0].coordinate.longitude)&client_id=\(Clientid)&client_secret=\(CLIENTSECRET)&v=20190520")  else {
-               return
-           }
-           
-        let request = URLRequest.init(url: url)
-        let task: URLSessionDataTask = session.dataTask(with: request){ (data, response, error) in
-            
-            guard let data = data else{
-            print(response ?? "no resonpse")
-            print(error ?? " no error")
-                return
-            }
-            let datacheck = String(decoding: data, as : UTF8.self)
-            print(datacheck)
-         }
-        task.resume()
-        
-        
-    }
-
-    
-    func locationManager(_ manager: CLLocationManager, didFailWithError error: Error) {
-        print("Location update falied , \(error)")
-    }
-    
 
 }
