+++ Dungeons and Dieting/Dungeons and Dieting/GoogleService-Info.plist
@@ -2,6 +2,8 @@
 <!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
 <plist version="1.0">
 <dict>
+	<key>FIREBASE_ANALYTICS_COLLECTION_DEACTIVATE</key>
+	<false/>
 	<key>CLIENT_ID</key>
 	<string>830669651170-0fh7fckeifrf3s7qukamdhbuu5g1t9ct.apps.googleusercontent.com</string>
 	<key>REVERSED_CLIENT_ID</key>
@@ -21,7 +23,7 @@
 	<key>IS_ADS_ENABLED</key>
 	<false/>
 	<key>IS_ANALYTICS_ENABLED</key>
-	<false/>
+	<true/>
 	<key>IS_APPINVITE_ENABLED</key>
 	<true/>
 	<key>IS_GCM_ENABLED</key>
