+++ Dungeons and Dieting/Dungeons and Dieting/MapViewController.swift
@@ -13,18 +13,29 @@ import GameplayKit
 import MapKit
 import CoreLocation
 
-class MapViewController: UIViewController , CLLocationManagerDelegate {
+class MapViewController: UIViewController , CLLocationManagerDelegate, MKMapViewDelegate {
    
  
-    
+    //buttons
     var friendButton: UIButton
     var itemButton: UIButton
     var mapButton: UIButton
     var progressButton: UIButton
     var shopButton: UIButton
-     var mapview: MKMapView
+      
     
+    
+    //map info
+    var mapview: MKMapView
+    let LocationM = CLLocationManager()
+    let Clientid = "WP0KU1PZ5EZELDOTL2MVVWEK0Y4PSU45TMHLHUCV4QFBK4OI"
+    let CLIENTSECRET = "FJJPDA5ANL3JCXAUALOMXKLWJYKSUY1PTZXESPGYO04RA1RK"
+     
+
+ var randomLoc = Double.random(in:-100...1000)
     init() {
+ 
+     
         mapview = MKMapView()
         friendButton = UIButton()
         itemButton = UIButton()
@@ -35,6 +46,7 @@ class MapViewController: UIViewController , CLLocationManagerDelegate {
     }
 
     required init?(coder aDecoder: NSCoder) {
+          
         mapview = MKMapView()
         friendButton = UIButton()
         itemButton = UIButton()
@@ -68,6 +80,15 @@ class MapViewController: UIViewController , CLLocationManagerDelegate {
         itemButton.addTarget(self, action: #selector(toItem), for: .touchUpInside)
         progressButton.addTarget(self, action: #selector(toProgress), for: .touchUpInside)
         shopButton.addTarget(self, action: #selector(toShop), for: .touchUpInside)
+                  setupLoation()
+        
+        
+
+             checkLocationServices()
+             self.mapview = MKMapView(frame: CGRect(x: 10, y: 60, width: 400, height: 600))
+
+              self.view.addSubview(mapview)
+          
     }
     
     @IBAction func toFriend() {
@@ -98,7 +119,112 @@ class MapViewController: UIViewController , CLLocationManagerDelegate {
         self.present(newViewController, animated: false, completion: nil)
     }
     
+    func checkLocationServices()
+     {
+         if(CLLocationManager.locationServicesEnabled())
+         {
+             setupLoation()
+             checkLocationAuthorization()
+         }
+     }
+   
     
+    ///marker info
+
+        
+
+    
+    func setupLoation()
+          {
+              LocationM.delegate = self
+              LocationM.desiredAccuracy = kCLLocationAccuracyHundredMeters
+              LocationM.startUpdatingLocation()
+          }
+    
+    func centerviewuserlocation(){
+        if let location = LocationM.location?.coordinate{
+            let region = MKCoordinateRegion.init(center: location, latitudinalMeters: 1000, longitudinalMeters: 1000)
+            mapview.setRegion(region, animated: true)
+            
+        }
+        
+    }
+
+    
+    
+    func checkLocationAuthorization() {
+        switch CLLocationManager.authorizationStatus() {
+        case .authorizedWhenInUse:
+            // map follws location
+            mapview.showsUserLocation = true;
+            centerviewuserlocation()
+        case.denied:
+            //show alert
+            LocationM.requestLocation()
+            break
+        case .notDetermined:
+            //plase turn on location services
+            LocationM.requestWhenInUseAuthorization()
+            break
+        case .restricted:
+            break
+        case .authorizedAlways:
+        break
+        default:
+            break
+        }
+        
+    }
+
+       
+       func locationManager(_ manager: CLLocationManager, didUpdateLocations locations: [CLLocation])
+       {
+           if let location =  locations.last
+           {
+               if location.horizontalAccuracy > 0 {
+                   LocationM.stopUpdatingLocation()
+                   let longitude =  location.coordinate.longitude
+                   let latitude  = location.coordinate.latitude
+                
+                let centerview = CLLocationCoordinate2D(latitude: latitude, longitude: longitude)
+                let region = MKCoordinateRegion.init(center: centerview, latitudinalMeters: 1000, longitudinalMeters: 1000)
+                mapview.setRegion(region, animated: true)
+               }
+           }
+           
+
+           let session = URLSession.shared
+           guard let url = URL(string: "https://api.foursquare.com/v2/venues/search?ll=\(locations[0].coordinate.latitude),\(locations[0].coordinate.longitude)&client_id=\(Clientid)&client_secret=\(CLIENTSECRET)&v=20190520")  else {
+                  return
+              }
+              
+           let request = URLRequest.init(url: url)
+           let task: URLSessionDataTask = session.dataTask(with: request){ (data, response, error) in
+               
+               guard let data = data else{
+               print(response ?? "no resonpse")
+               print(error ?? " no error")
+                   return
+               }
+               let datacheck = String(decoding: data, as : UTF8.self)
+               print(datacheck)
+            }
+           task.resume()
+           
+           
+       }
+
+       
+       func locationManager(_ manager: CLLocationManager, didFailWithError error: Error) {
+           print("Location update falied , \(error)")
+       }
+       
+    
+       
+    func locationManager(_ manager: CLLocationManager, didChangeAuthorization status: CLAuthorizationStatus)
+    {
+     checkLocationAuthorization()
+    }
 
 
 }
